requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
pinecone-client>=3.0.0,<4.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.1
langchain-openai>=0.0.5
langchain-text-splitters>=0.0.1
sentence-transformers>=5.0.0